fixed many to many join between services and bookings

+++ server/src/main/java/com/example/cleaner/repositories/CleanerRepository.java
@@ -3,5 +3,5 @@ package com.example.cleaner.repositories;
 import com.example.cleaner.models.users.Cleaner;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CleanerRepository extends JpaRepository<Cleaner, Integer> {
+public interface CleanerRepository extends JpaRepository<Cleaner, Long> {
 }
